move files into datasource
it's a better representation for what they are doing.
these files are datasources one reading from a file through ssh and the
other reading from docker
